backend/requirements.txt
aiofiles==23.2.1
alembic==1.17.0
annotated-types==0.7.0
anyio==4.10.0
asttokens==3.0.0
asyncpg==0.30.0
audioop-lts==0.2.2
audioread==3.0.1
backoff==2.2.1
bcrypt==4.3.0
BeatNet==1.1.1
black==23.11.0
boto3==1.40.60
botocore==1.40.61
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
contourpy==1.3.3
coverage==7.11.0
cryptography==46.0.3
cycler==0.12.1
Cython==3.1.4
decorator==5.2.1
Deprecated==1.2.18
distro==1.9.0
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.1.0
essentia==2.1b6.dev1389
executing==2.2.1
Faker==20.1.0
fastapi==0.119.1
ffmpeg-python==0.2.0
filelock==3.20.0
flake8==6.1.0
fonttools==4.60.1
fsspec==2025.9.0
future==1.0.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
inngest==0.5.11
ipython==8.18.1
jcs==0.2.1
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.2
kiwisolver==1.4.9
lazy_loader==0.4
librosa==0.11.0
limits==5.6.0
llvmlite==0.45.1
madmom==0.16.1
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib==3.10.7
matplotlib-inline==0.2.1
mccabe==0.7.0
mido==1.3.3
mpmath==1.3.0
msgpack==1.1.2
mypy==1.7.1
mypy_extensions==1.1.0
networkx==3.5
numba==0.62.1
numpy==2.2.6
packaging==25.0
parso==0.8.5
passlib==1.7.4
pathspec==0.12.1
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.5.0
pluggy==1.6.0
pooch==1.8.2
posthog==6.9.2
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pycodestyle==2.11.1
pycparser==2.23
pydantic==2.12.2
pydantic-settings==2.1.0
pydantic_core==2.41.4
pyflakes==3.1.0
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.5
pytest==8.4.2
pytest-asyncio==0.21.1
pytest-cov==4.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-jose==3.3.0
python-magic==0.4.27
python-multipart==0.0.6
PyYAML==6.0.2
requests==2.32.5
rsa==4.9.1
s3transfer==0.14.0
scikit-learn==1.7.2
scipy==1.14.1
setuptools==80.9.0
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
soundfile==0.13.1
soxr==1.0.0
SQLAlchemy==2.0.44
stack-data==0.6.3
standard-aifc==3.13.0
standard-chunk==3.13.0
standard-sunau==3.13.0
starlette==0.48.0
stripe==11.4.0
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.8.0
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.21.0
watchfiles==1.1.0
wcwidth==0.2.14
websockets==15.0.1
wheel==0.45.1
wrapt==1.17.3
yt-dlp==2025.9.5
postmarker==1.0
